[Feature] refactored the task logic - now the task components are defined by what blocks are available and the exact configuration of ball and robot are set in task.json
