+++ src/tasks/Kick.js
@@ -5,7 +5,7 @@ import {useBlockly} from "../helper/useBlockly";
 import RobotActions from "../robocup/RobotActions";
 import {useDispatch} from "react-redux";
 import BallActions from "../robocup/BallActions";
-import ExecuteResetButton from "./ExecuteResetButton";
+import ExecuteResetButton from "../helper/ExecuteResetButton";
 
 /**
  * TASK
@@ -19,7 +19,7 @@ import ExecuteResetButton from "./ExecuteResetButton";
  * @returns {*}
  * @constructor
  */
-const GoalKick = () => {
+const Kick = ({task_properties}) => {
 
     const blockly = useBlockly();
 
@@ -28,9 +28,9 @@ const GoalKick = () => {
     // Initialize the robot position on the field for the given task
     const reset = () => {
         dispatch(RobotActions.reset());
-        dispatch(RobotActions.addRobot(460,220, 90));
+        dispatch(RobotActions.addRobot(task_properties.own_robot.position.x,task_properties.own_robot.position.y, task_properties.own_robot.position.rotation));
         dispatch(BallActions.updateBall(0,0));
-        dispatch(BallActions.moveBall(470,220));
+        dispatch(BallActions.moveBall(task_properties.ball.position.x,task_properties.ball.position.y));
     };
 
     useEffect(reset, []);
@@ -38,10 +38,7 @@ const GoalKick = () => {
     return(
         <div>
             <div>
-                <RoboCupField
-                    grid_properties={{
-                    }}
-                />
+                <RoboCupField/>
             </div>
             <ExecuteResetButton execute={blockly.generateCode} reset={reset} />
             <BlocklyComponent ref={blockly.simpleWorkspace}
@@ -59,4 +56,4 @@ const GoalKick = () => {
 
 };
 
-export default GoalKick
+export default Kick
