+++ src/tasks/Walk.js
@@ -2,24 +2,24 @@ import {RoboCupField} from "../robocup/field";
 import BlocklyComponent, {Block, Field, Shadow, Value} from "../Blockly";
 import React, {useEffect} from "react";
 import {useBlockly} from "../helper/useBlockly";
-import {useDispatch} from "react-redux";
 import RobotActions from "../robocup/RobotActions";
+import {useDispatch} from "react-redux";
 import BallActions from "../robocup/BallActions";
-import ExecuteResetButton from "./ExecuteResetButton";
+import ExecuteResetButton from "../helper/ExecuteResetButton";
 
 /**
  * TASK
  * ====
  *
- * Initial: Ball positioned on right penalty point, robot one field below center mark, facing the center mark
+ * Initial: Ball positioned on right penalty point, robot three fields away from the ball, facing the goal
  * Task: Kick ball into goal
- * Required Actions: Turn, Walk, Kick
+ * Required Actions: Walk, Kick
  * Required Coding Concepts: --
  *
  * @returns {*}
  * @constructor
  */
-const StraightAndTurn = () => {
+const Walk = ({task_properties}) => {
 
     const blockly = useBlockly();
 
@@ -28,20 +28,17 @@ const StraightAndTurn = () => {
     // Initialize the robot position on the field for the given task
     const reset = () => {
         dispatch(RobotActions.reset());
-        dispatch(RobotActions.addRobot(300,250, 0));
+        dispatch(RobotActions.addRobot(task_properties.own_robot.position.x,task_properties.own_robot.position.y, task_properties.own_robot.position.rotation));
         dispatch(BallActions.updateBall(0,0));
-        dispatch(BallActions.moveBall(470,220));
-    }
+        dispatch(BallActions.moveBall(task_properties.ball.position.x,task_properties.ball.position.y));
+    };
 
     useEffect(reset, []);
 
     return(
         <div>
             <div>
-                <RoboCupField
-                    grid_properties={{
-                    }}
-                />
+                <RoboCupField/>
             </div>
             <ExecuteResetButton execute={blockly.generateCode} reset={reset} />
             <BlocklyComponent ref={blockly.simpleWorkspace}
@@ -53,8 +50,6 @@ const StraightAndTurn = () => {
                               }}>
                 <Block type="move_one_block_ahead"/>
                 <Block type="ball_kick"/>
-                <Block type="turn_right"/>
-                <Block type="turn_left"/>
             </BlocklyComponent>
         </div>
     )
@@ -62,4 +57,4 @@ const StraightAndTurn = () => {
 
 };
 
-export default StraightAndTurn
+export default Walk
